cybersec_lab/requirements.txt
streamlit
streamlit-authenticator
anthropic
python-dotenv
pandas
plotly
markupsafe
